hiding the RawModule type from main
